# Conflicts:
#	src/main/java/devut/buzzerbidder/domain/user/service/UserService.java
#	src/main/java/devut/buzzerbidder/global/exeption/ErrorCode.java

+++ src/main/java/devut/buzzerbidder/domain/liveitem/dto/request/LiveItemSearchRequest.java
@@ -5,7 +5,7 @@ import devut.buzzerbidder.domain.liveitem.entity.LiveItem.Category;
 public record LiveItemSearchRequest(
     String name,
     Category category,
-    Integer minBidPrice,
-    Integer maxBidPrice
+    Long minBidPrice,
+    Long maxBidPrice
 ){
 }
